+++ src/tribler-core/tribler_core/components/gigachannel_manager/tests/test_gigachannel_manager_component.py
@@ -1,0 +1,26 @@
+from unittest.mock import patch
+
+from tribler_core.components.base import Session
+from tribler_core.components.gigachannel_manager.gigachannel_manager_component import GigachannelManagerComponent
+from tribler_core.components.libtorrent import LibtorrentComponent
+from tribler_core.components.masterkey import MasterKeyComponent
+from tribler_core.components.metadata_store.metadata_store_component import MetadataStoreComponent
+from tribler_core.components.restapi import RESTComponent
+from tribler_core.components.socks_configurator import SocksServersComponent
+# pylint: disable=protected-access
+from tribler_core.restapi.rest_manager import RESTManager
+
+
+async def test_gigachannel_manager_component(tribler_config):
+    components = [SocksServersComponent(), MasterKeyComponent(), RESTComponent(), MetadataStoreComponent(),
+                  LibtorrentComponent(), GigachannelManagerComponent()]
+    session = Session(tribler_config, components)
+    with session:
+        comp = GigachannelManagerComponent.instance()
+        with patch.object(RESTManager, 'get_endpoint'):
+            await session.start()
+
+            assert comp.gigachannel_manager
+            assert comp._rest_manager
+
+            await session.shutdown()
